+++ Changes.md
@@ -1,3 +1,8 @@
+Change Log in version 401.1.2 (2022112303)
+===========================================
+1. Fix 'blockside' setting not being applied to the course index and side post.
+2. Fix 'No nobile navigation' on navbar.
+
 Change Log in version 401.1.1 (2022112302)
 ===========================================
 1. Layout style tidy.